Move the CustomExtension decorator from the v21.observables module to v21.common.  Custom extensions are not specific to SCOs, so I don't know why it was in that module.  Now, ExtensionDefinition and CustomExtension are together in the same module, just like MarkingDefinition and CustomMarking are together.  Made sense to me.
